Rearrange file position and remove .h in plugins/
According to: https://github.com/cms-sw/cmssw/pull/38226#issuecomment-1500148966
